ingestion/pyproject.toml
[project]
name = "az-scout-bdd-sku-ingestion"
version = "0.1.0"
description = "Ingestion CLI for az-scout-plugin-bdd-sku – writes VM pricing data into PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "requests>=2.32",
    "azure-identity>=1.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
